test: move helper coverage into extensions
