+++ extensions/amazon-bedrock/discovery.test.ts
@@ -1,5 +1,11 @@
 import type { BedrockClient } from "@aws-sdk/client-bedrock";
 import { beforeEach, describe, expect, it, vi } from "vitest";
+import {
+  discoverBedrockModels,
+  mergeImplicitBedrockProvider,
+  resetBedrockDiscoveryCacheForTest,
+  resolveBedrockConfigApiKey,
+} from "./api.js";
 
 const sendMock = vi.fn();
 const clientFactory = () => ({ send: sendMock }) as unknown as BedrockClient;
@@ -14,12 +20,6 @@ const baseActiveAnthropicSummary = {
   modelLifecycle: { status: "ACTIVE" },
 };
 
-async function loadDiscovery() {
-  const mod = await import("../plugin-sdk/amazon-bedrock.js");
-  mod.resetBedrockDiscoveryCacheForTest();
-  return mod;
-}
-
 function mockSingleActiveSummary(overrides: Partial<typeof baseActiveAnthropicSummary> = {}): void {
   sendMock.mockResolvedValueOnce({
     modelSummaries: [{ ...baseActiveAnthropicSummary, ...overrides }],
@@ -29,11 +29,10 @@ function mockSingleActiveSummary(overrides: Partial<typeof baseActiveAnthropicSu
 describe("bedrock discovery", () => {
   beforeEach(() => {
     sendMock.mockClear();
+    resetBedrockDiscoveryCacheForTest();
   });
 
   it("filters to active streaming text models and maps modalities", async () => {
-    const { discoverBedrockModels } = await loadDiscovery();
-
     sendMock.mockResolvedValueOnce({
       modelSummaries: [
         {
@@ -88,7 +87,6 @@ describe("bedrock discovery", () => {
   });
 
   it("applies provider filter", async () => {
-    const { discoverBedrockModels } = await loadDiscovery();
     mockSingleActiveSummary();
 
     const models = await discoverBedrockModels({
@@ -100,7 +98,6 @@ describe("bedrock discovery", () => {
   });
 
   it("uses configured defaults for context and max tokens", async () => {
-    const { discoverBedrockModels } = await loadDiscovery();
     mockSingleActiveSummary();
 
     const models = await discoverBedrockModels({
@@ -112,7 +109,6 @@ describe("bedrock discovery", () => {
   });
 
   it("caches results when refreshInterval is enabled", async () => {
-    const { discoverBedrockModels } = await loadDiscovery();
     mockSingleActiveSummary();
 
     await discoverBedrockModels({ region: "us-east-1", clientFactory });
@@ -121,8 +117,6 @@ describe("bedrock discovery", () => {
   });
 
   it("skips cache when refreshInterval is 0", async () => {
-    const { discoverBedrockModels } = await loadDiscovery();
-
     sendMock
       .mockResolvedValueOnce({ modelSummaries: [baseActiveAnthropicSummary] })
       .mockResolvedValueOnce({ modelSummaries: [baseActiveAnthropicSummary] });
@@ -140,9 +134,7 @@ describe("bedrock discovery", () => {
     expect(sendMock).toHaveBeenCalledTimes(2);
   });
 
-  it("resolves the Bedrock config apiKey from AWS auth env vars", async () => {
-    const { resolveBedrockConfigApiKey } = await loadDiscovery();
-
+  it("resolves the Bedrock config apiKey from AWS auth env vars", () => {
     expect(
       resolveBedrockConfigApiKey({
         AWS_BEARER_TOKEN_BEDROCK: "bearer", // pragma: allowlist secret
@@ -153,9 +145,7 @@ describe("bedrock discovery", () => {
     expect(resolveBedrockConfigApiKey({} as NodeJS.ProcessEnv)).toBe("AWS_PROFILE");
   });
 
-  it("merges implicit Bedrock models into explicit provider overrides", async () => {
-    const { mergeImplicitBedrockProvider } = await loadDiscovery();
-
+  it("merges implicit Bedrock models into explicit provider overrides", () => {
     expect(
       mergeImplicitBedrockProvider({
         existing: {
@@ -180,91 +170,6 @@ describe("bedrock discovery", () => {
           ],
         },
       }).models?.map((model) => model.id),
-    ).toEqual(["amazon.nova-micro-v1:0"]);
-  });
-
-  it("merges implicit Bedrock discovery into provider catalog config", async () => {
-    vi.resetModules();
-    const bedrockApi = await import("../plugin-sdk/amazon-bedrock.js");
-    vi.spyOn(bedrockApi, "resolveImplicitBedrockProvider").mockResolvedValue({
-      baseUrl: "https://bedrock-runtime.us-east-1.amazonaws.com",
-      api: "bedrock-converse-stream",
-      auth: "aws-sdk",
-      models: [
-        {
-          id: "amazon.nova-micro-v1:0",
-          name: "Nova Micro",
-          reasoning: false,
-          input: ["text"],
-          cost: { input: 0, output: 0, cacheRead: 0, cacheWrite: 0 },
-          contextWindow: 32_000,
-          maxTokens: 4096,
-        },
-      ],
-    });
-    vi.spyOn(bedrockApi, "mergeImplicitBedrockProvider").mockImplementation(
-      ({ existing, implicit }) => ({
-        ...implicit,
-        ...existing,
-        models:
-          Array.isArray(existing?.models) && existing.models.length > 0
-            ? existing.models
-            : implicit.models,
-      }),
-    );
-    const result = await (async () => {
-      const implicit = await bedrockApi.resolveImplicitBedrockProvider({
-        config: {
-          models: {
-            bedrockDiscovery: {
-              enabled: true,
-            },
-          },
-        },
-        env: {
-          AWS_PROFILE: "default",
-        } as NodeJS.ProcessEnv,
-      });
-      if (!implicit) {
-        return null;
-      }
-      return {
-        provider: bedrockApi.mergeImplicitBedrockProvider({
-          existing: {
-            baseUrl: "https://bedrock-runtime.us-west-2.amazonaws.com",
-            headers: { "x-test-header": "1" },
-            models: [],
-          },
-          implicit,
-        }),
-      };
-    })();
-
-    expect(bedrockApi.resolveImplicitBedrockProvider).toHaveBeenCalledWith({
-      config: {
-        models: {
-          bedrockDiscovery: {
-            enabled: true,
-          },
-        },
-      },
-      env: {
-        AWS_PROFILE: "default",
-      } as NodeJS.ProcessEnv,
-    });
-
-    expect(result).toMatchObject({
-      provider: {
-        api: "bedrock-converse-stream",
-        auth: "aws-sdk",
-        baseUrl: "https://bedrock-runtime.us-west-2.amazonaws.com",
-        headers: { "x-test-header": "1" },
-      },
-    });
-    expect(
-      result && "provider" in result
-        ? result.provider.models?.map((model: { id: string }) => model.id)
-        : [],
     ).toEqual(["amazon.nova-micro-v1:0"]);
   });
 });
